pyproject.toml
[tool.poetry]
name = "whats-hard-llm-powered-apps"
version = "0.1.0"
description = ""
authors = ["Ian Massingham <ian@griptape.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
griptape = {extras = ["all"], version = "^1.3.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
